+++ SharedAssemblyInfo.cs
@@ -2,8 +2,8 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyCompany("Brian Lehnen")]
-[assembly: AssemblyCopyright("Copyright © Brian Lehnen 2023")]
+[assembly: AssemblyCopyright("Copyright © Brian Lehnen 2024")]
 [assembly: ComVisible(false)]
-[assembly: AssemblyVersion("0.7.3")]
-[assembly: AssemblyFileVersion("0.7.3")]
-[assembly: AssemblyInformationalVersion("0.7.3")]
+[assembly: AssemblyVersion("0.7.4")]
+[assembly: AssemblyFileVersion("0.7.4")]
+[assembly: AssemblyInformationalVersion("0.7.4")]