+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+###0.7.4 Future
+* Add test for scheduler creation with memory queue
+* Move two logging messages from info -> debug
+* Add net8.0 as target
+
 ###0.7.3 2023-11-28
 * Error notification will not happen if a rollback notification is being performed